Fix exercice 4

+++ applications/self-star/temperatureManagement/exerciceFour/roomOccupancy/src/main/java/fr/liglab/icasa/self/star/temperature/management/exercice/four/room/occupancy/RoomOccupancy.java
@@ -29,6 +29,6 @@ public interface RoomOccupancy {
      *         that there the room is always empty and 1 indicates that the room
      *         is always occupied at the given moment of the day.
      */
-    public double getRoomOccupancy(String room,double minuteOfTheDay, String user);
+    public double getRoomOccupancy(String room,int minuteOfTheDay, String user);
 
 }
